requirements.txt
flask==2.3.3
flask-login==0.6.3
python-dateutil==2.8.2
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
pytest
pytest-flask
pytest-cov